requirements-dev.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.4
lxml==4.4.1
python-decouple==3.1
python-docx==0.8.10
pytz==2019.2
sqlparse==0.3.0
static3==0.7.0
whitenoise==4.1.4